+++ core/src/detect-ad-blocker.ts
