+++ core/src/event-timer.spec.ts
