+++ core/src/event-timer.ts
