+++ core/src/types.ts
@@ -1,0 +1,15 @@
+export type ConnectionType =
+	| 'bluetooth'
+	| 'cellular'
+	| 'ethernet'
+	| 'mixed'
+	| 'none'
+	| 'other'
+	| 'unknown'
+	| 'wifi';
+
+export interface NetworkInformation extends EventTarget {
+	readonly type?: ConnectionType;
+	readonly downlink?: number;
+	readonly effectiveType?: string;
+}
