+++ core/tsconfig.json
@@ -6,8 +6,8 @@
 		"outDir": "dist",
 		"typeRoots": ["./node_modules/@types", "./node_modules/web-vitals/dist"]
 	},
-	"extends": "./tsconfig.json",
-	"rootDir": "src",
+	"extends": "../tsconfig.json",
+	"rootDir": "./src",
 	"include": ["src/export.ts", "src/types"],
 	"exclude": ["**/*.spec.*", "**/*.test.*"]
 }
