+++ package.json
@@ -1,5 +1,5 @@
 {
-	"name": "@guardian/commercial",
+	"name": "@guardian/commercial-monorepo",
 	"version": "26.1.2",
 	"description": "Guardian advertising business logic",
 	"homepage": "https://github.com/guardian/commercial#readme",
@@ -14,15 +14,11 @@
 	"files": [
 		"dist"
 	],
-	"main": "dist/core/cjs/export.js",
-	"module": "dist/core/esm/export.js",
 	"scripts": {
-		"build": "npm-run-all clean --parallel compile:core:* build:prod build:dev",
+		"build": "npm-run-all clean --parallel build:*",
 		"build:dev": "webpack -c webpack.config.dev.mjs",
 		"build:prod": "webpack -c webpack.config.prod.mjs",
 		"clean": "rm -rf dist",
-		"compile:core:common": "tsc --project ./tsconfig.core.json --outDir ./dist/core/cjs --module CommonJS",
-		"compile:core:esm": "tsc --project ./tsconfig.core.json --outDir ./dist/core/esm",
 		"playwright:run": "playwright test",
 		"playwright:open": "playwright test --ui",
 		"lint": "eslint .",
@@ -47,6 +43,7 @@
 	"dependencies": {
 		"@guardian/ab-core": "8.0.1",
 		"@guardian/core-web-vitals": "11.0.0",
+		"@guardian/commercial": "workspace:*",
 		"@guardian/identity-auth": "^7.0.0",
 		"@guardian/identity-auth-frontend": "9.0.0",
 		"@guardian/libs": "22.5.0",
