+++ src/experiments/ab.ts
@@ -1,5 +1,5 @@
-import type { ABTest } from '@guardian/ab-core';
 import { AB } from '@guardian/ab-core';
+import type { ABTest } from '@guardian/ab-core';
 import type { OphanRecordFunction } from '@guardian/ab-core/dist/@types';
 import { getCookie, log } from '@guardian/libs';
 import {
