+++ src/lib/ad-verification/prepare-ad-verification.ts
@@ -1,6 +1,6 @@
+import { EventTimer } from '@guardian/commercial/event-timer';
 import { loadScript, log } from '@guardian/libs';
 import { refreshAdvert } from '../../display/load-advert';
-import { EventTimer } from '../../lib/event-timer';
 import { bypassCommercialMetricsSampling } from '../../lib/send-commercial-metrics';
 import type { ConfiantCallback } from '../../types/global';
 import { getAdvertById } from '../dfp/get-advert-by-id';
