+++ src/lib/commercial-boot-utils.ts
@@ -1,7 +1,7 @@
+import { EventTimer } from '@guardian/commercial/event-timer';
 import { log } from '@guardian/libs';
 import { adSlotIdPrefix } from './dfp/dfp-env-globals';
 import { reportError } from './error/report-error';
-import { EventTimer } from './event-timer';
 
 const tags: Record<string, string> = {
 	bundle: 'standalone',
