+++ src/lib/send-commercial-metrics.spec.ts
@@ -1,6 +1,6 @@
+import { EventTimer } from '@guardian/commercial/event-timer';
 import type { ConsentState, USNATConsentState } from '@guardian/libs';
 import { onConsent } from '@guardian/libs';
-import { EventTimer } from './event-timer';
 import {
 	_,
 	bypassCommercialMetricsSampling,
