+++ src/lib/track-gpc-signal.spec.ts
@@ -1,5 +1,5 @@
+import { EventTimer } from '@guardian/commercial/event-timer';
 import type { ConsentState, USNATConsentState } from '@guardian/libs';
-import { EventTimer } from './event-timer';
 import { initTrackGpcSignal } from './track-gpc-signal';
 
 describe('initTrackGpcSignal', () => {
