+++ src/lib/track-scroll-depth.spec.ts
@@ -1,4 +1,4 @@
-import { EventTimer } from './event-timer';
+import { EventTimer } from '@guardian/commercial/event-timer';
 import { initTrackScrollDepth } from './track-scroll-depth';
 
 describe('initTrackScrollDepth', () => {
