+++ src/lib/track-scroll-depth.ts
@@ -1,5 +1,5 @@
+import { EventTimer } from '@guardian/commercial/event-timer';
 import { log } from '@guardian/libs';
-import { EventTimer } from './event-timer';
 
 /**
  * Collect commercial metrics on scroll depth
