+++ src/lib/types.ts
@@ -1,21 +1,5 @@
 import type { VendorName } from '@guardian/libs';
 
-export type ConnectionType =
-	| 'bluetooth'
-	| 'cellular'
-	| 'ethernet'
-	| 'mixed'
-	| 'none'
-	| 'other'
-	| 'unknown'
-	| 'wifi';
-
-export interface NetworkInformation extends EventTarget {
-	readonly type?: ConnectionType;
-	readonly downlink?: number;
-	readonly effectiveType?: string;
-}
-
 export type TagAttribute = {
 	name: string;
 	value: string;
