+++ src/types/global.ts
@@ -1,12 +1,12 @@
+import type { EventTimer } from '@guardian/commercial/event-timer';
+import type { NetworkInformation } from '@guardian/commercial/types';
 import type { CustomClaims } from '@guardian/identity-auth';
 import type { VendorName } from '@guardian/libs';
 import type { DfpEnv } from '../lib/dfp/dfp-env';
-import type { EventTimer } from '../lib/event-timer';
 import type { PageTargeting } from '../lib/targeting/build-page-targeting';
 import type {
 	GoogleTagParams,
 	GoogleTrackConversionObject,
-	NetworkInformation,
 } from '../lib/types';
 import type { IasPETSlot } from './ias';
